estoque0.5

+++ Projeto_Estoque/sistema.py
@@ -1,6 +1,10 @@
+import sqlite3
 import sys
+
+from PySide2.QtCore import Qt
+
 from bd_sqlite3 import DataBase
-from PySide2.QtWidgets import (QApplication, QMainWindow,QWidget,QMessageBox)
+from PySide2.QtWidgets import (QApplication, QMainWindow, QWidget, QMessageBox, QTableWidgetItem, QDialog, QInputDialog)
 from PySide2.QtSql import QSqlDatabase, QSqlTableModel
 from ui_login import Ui_Login
 from ui_MainWindow import Ui_MainWindow
@@ -26,86 +30,26 @@ class Login(QWidget, Ui_Login):
 
     #Verifica a senha e chama o Sistema caso True
     def verifica_senha(self):
-        try:
-            db = DataBase()
-            db.conecta()
-            if db.db_check_user_admin(self.lineEdit_usuario.text().strip(),self.lineEdit_senha.text().strip()) == 'user':
-                self.login = self.lineEdit_usuario.text()
-                self.senha = self.lineEdit_senha.text()
-                self.w = MainWindow(self.login,self.senha)
-                self.w.bt_novo_usuario.hide()
-                self.w.show()
-                self.close()
-                db.close_conecta()
-            elif db.db_check_user_admin(self.lineEdit_usuario.text().strip(),self.lineEdit_senha.text().strip()) == 'admin':
-                self.login = self.lineEdit_usuario.text()
-                self.senha = self.lineEdit_senha.text()
-                self.w = MainWindow(self.login,self.senha)
-                self.w.show()
-                self.close()
-                db.close_conecta()
-        except:
-                    # abre uma caixa de msg com erro
+        db = DataBase()
+        db.conecta()
+        if db.db_check_user_admin(self.lineEdit_usuario.text().strip(),self.lineEdit_senha.text().strip()) == 'user':
+            self.login = self.lineEdit_usuario.text()
+            self.senha = self.lineEdit_senha.text()
+            self.w = MainWindow(self.login,self.senha)
+            self.w.bt_novo_usuario.hide()
+            self.w.show()
+            self.close()
+            db.close_conecta()
+        elif db.db_check_user_admin(self.lineEdit_usuario.text().strip(),self.lineEdit_senha.text().strip()) == 'admin':
+            self.login = self.lineEdit_usuario.text()
+            self.senha = self.lineEdit_senha.text()
+            self.w = MainWindow(self.login,self.senha)
+            self.w.show()
+            self.close()
+            db.close_conecta()
+        else:
             self.messagebox_critical('LOGIN OU SENHA INVÁLIDOS')
             self.lineEdit_senha.setText('')
-
-
-class Produtos(QWidget, Ui_Produtos):
-    def __init__(self,user) -> None:
-        super(Produtos,self).__init__()
-        self.setupUi(self)
-        self.user = user
-
-        self.bt_novo_produto.clicked.connect(self.add_produto)
-
-    def messagebox_critical(self, txt):
-        msg = QMessageBox()
-        msg.setIcon(msg.Critical)
-        msg.setWindowTitle('ERRO DE DADOS')
-        msg.setText(txt)
-        msg.exec()
-
-    def messagebox_accept(self, txt):
-        msg = QMessageBox()
-        msg.setIcon(QMessageBox.Information)
-        msg.setWindowTitle("DADOS ACEITOS")
-        msg.setText(txt)
-        msg.exec_()
-
-
-    def add_produto(self):
-        produto = self.le_produto.text()
-        user = self.user
-        ref = self.comboBox_quantidade.currentText()
-        quantidade = self.le_quantidade.text()
-        quantidade_int = int(quantidade)
-        valor = self.le_valor.text().replace(',','.')
-        valor_int = float(valor)
-        valor_total = f"{valor_int*quantidade_int:.2f}".replace('.',',')
-        valorstr = self.le_valor.text()
-        if ref == 'KG':
-            db = DataBase()
-            db.conecta()
-            db.insert_novo_produto(produto,user,valorstr,valor_total,kg=quantidade)
-            self.messagebox_accept('PRODUTO CADASTRADO COM SUCESSO')
-            self.close()
-            db.close_conecta()
-        elif ref == 'G':
-            db = DataBase()
-            db.conecta()
-            db.insert_novo_produto(produto, user, valorstr, valor_total, g=quantidade)
-            self.messagebox_accept('PRODUTO CADASTRADO COM SUCESSO')
-            self.close()
-            db.close_conecta()
-        elif ref == 'UN':
-            db = DataBase()
-            db.conecta()
-            db.insert_novo_produto(produto, user, valorstr, valor_total, un=quantidade)
-            self.messagebox_accept('PRODUTO CADASTRADO COM SUCESSO')
-            self.close()
-            db.close_conecta()
-
-
 
 # instancia do Sistema
 class MainWindow(QMainWindow, Ui_MainWindow):
@@ -128,7 +72,6 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
                 #BOTAO PARA CADASTRAR CLIENTE NO BANCO DE DADOS
         self.bt_cadastrar.clicked.connect(self.novo_cliente)
-
         self.bt_cadastrar_usuario.clicked.connect(self.new_user)
         self.bt_deletar_usuario.clicked.connect(self.delete_user)
 
@@ -138,6 +81,40 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         self.bt_importar.clicked.connect(self.import_produto)
 
         self.reset_tables()
+
+        self.tv_estoque.doubleClicked.connect(self.get_produto_via_table)
+        # self.bt_gerar_saida.clicked.connect(self.gerar_saida)
+        self.le_valor.textChanged.connect(self.mascara_valor)
+        self.le_quantidade.textChanged.connect(self.mascara_quantidade)
+        self.le_valor.textEdited.connect(self.clear_id)
+        self.le_quantidade.textEdited.connect(self.clear_id)
+        self.le_produto.textEdited.connect(self.clear_id)
+
+        self.le_valor_saida.setReadOnly(True)
+        self.le_produto_saida.setReadOnly(True)
+        self.le_ID_produto_saida.setReadOnly(True)
+        self.le_ID_produto.setReadOnly(True)
+        self.comboBox_un_saida.setEnabled(False)
+
+
+    def mascara_valor(self,s):
+        l = [i for i in s if i.isdigit() or i == ',' or i == '.']
+        p = ''
+        for i in l:
+            p += str(i).strip()
+        self.le_valor.setText(f'R${p}')
+
+    def mascara_quantidade(self,s):
+        l = [i for i in s if i.isdigit()]
+        p = ''
+        for i in l:
+            p += str(i).strip()
+        self.le_quantidade.setText(f'{p}')
+
+    def clear_id(self):
+        self.le_ID_produto.clear()
+
+
 
     def get_usuario(self,login,senha):
         try:
@@ -150,31 +127,112 @@ class MainWindow(QMainWindow, Ui_MainWindow):
             print(self.usuario)
             db.close_conecta()
         except:
-            self.messagebox_critical('Usuario não encontrado')
+            self.messagebox_critical('Usuario não encontrado','Usuario não encontrado')
 
     def import_produto(self):
-        self.p = Produtos(self.usuario)
-        self.p.show()
-        self.reset_tables()
-
-
-
-
-
+        produto = self.le_produto.text()
+        user = self.usuario
+        ref = self.comboBox_un.currentText()
+        quantidade = self.le_quantidade.text()
+        quantidade_int = int(quantidade)
+        valor = self.le_valor.text().replace(',', '.').replace('R','').replace('$','')
+        valor_int = float(valor)
+        valor_total = f"{valor_int * quantidade_int:.2f}".replace('.', ',')
+        valorstr = self.le_valor.text().replace(',', '.').replace('R','').replace('$','')
+        if ref == 'KG':
+            db = DataBase()
+            db.conecta()
+            db.insert_novo_produto(produto, user, valorstr, valor_total, kg=quantidade)
+            self.messagebox_accept('PRODUTO CADASTRADO','PRODUTO CADASTRADO COM SUCESSO')
+            self.reset_tables()
+            self.le_produto.setText('')
+            self.le_valor.setText('')
+            self.le_quantidade.setText('')
+            db.close_conecta()
+        elif ref == 'G':
+            db = DataBase()
+            db.conecta()
+            db.insert_novo_produto(produto, user, valorstr, valor_total, g=quantidade)
+            self.messagebox_accept('PRODUTO CADASTRADO','PRODUTO CADASTRADO COM SUCESSO')
+            self.reset_tables()
+            self.le_produto.setText('')
+            self.le_valor.setText('')
+            self.le_quantidade.setText('')
+            db.close_conecta()
+        elif ref == 'UN':
+            db = DataBase()
+            db.conecta()
+            db.insert_novo_produto(produto, user, valorstr, valor_total, un=quantidade)
+            self.messagebox_accept('PRODUTO CADASTRADO','PRODUTO CADASTRADO COM SUCESSO')
+            self.reset_tables()
+            self.le_produto.setText('')
+            self.le_valor.setText('')
+            self.le_quantidade.setText('')
+            db.close_conecta()
 
         #FUNCAO QUE VERIFICA SE AS SENHAS INFORMADAS AO CRIAR UM NOVO USUARIO ESTAO VAZIAS OU NAO SÃO IGUAIS
 
-    def messagebox_critical(self, txt):
-        msg = QMessageBox()
+    def get_produto_via_table(self):
+        y = self.tv_estoque.currentIndex()
+        x = self.model_estoque.primaryValues(y.row())
+        id_produto = x.value('ID_produto')
+        db = DataBase()
+        db.conecta()
+        cursor = db.conection.cursor()
+        cursor.execute(f'select * from estoque where ID_PRODUTO = "{id_produto}";')
+        for campo in cursor.fetchall():
+            id_produto = campo[0]
+            produto = campo[1]
+            valor = campo[2].replace('R','').replace('$','')
+            quantidade_kg = campo[5]
+            quantidade_g = campo[6]
+            quantidade_un = campo[4]
+            ref = '--'
+            if quantidade_un == ref and quantidade_g == ref:
+                self.comboBox_un.setCurrentIndex(0)
+                quantidade = quantidade_kg
+            elif quantidade_kg == ref and quantidade_g == ref:
+                self.comboBox_un.setCurrentIndex(2)
+                quantidade = quantidade_un
+            else:
+                self.comboBox_un.setCurrentIndex(1)
+                quantidade = quantidade_g
+            self.le_ID_produto.setText(str(id_produto))
+            self.le_produto.setText(produto)
+            self.le_valor.setText(valor)
+            self.le_quantidade.setText(quantidade)
+
+            db.close_conecta()
+
+    # def gerar_saida(self):
+    #     db = DataBase()
+    #     db.conecta()
+    #     cursor = db.conection.cursor()
+    #     id_produto = self.le_ID_produto.text()
+    #     cursor.execute(f'select * from estoque where ID_PRODUTO = "{id_produto}";')
+    #     for campo in cursor.fetchall():
+
+
+
+
+    def messagebox_critical(self,title, txt):
+        msg = QMessageBox(self)
         msg.setIcon(msg.Critical)
-        msg.setWindowTitle('ERRO DE DADOS')
+        msg.setWindowTitle(title)
         msg.setText(txt)
         msg.exec()
 
-    def messagebox_accept(self, txt):
-        msg = QMessageBox()
+    def messagebox_aviso(self,title, txt):
+        msg = QMessageBox(self)
+        msg.setIcon(msg.Warning)
+        msg.setWindowTitle(title)
+        msg.setText(txt)
+        msg.exec()
+
+    def messagebox_accept(self,title, txt):
+        msg = QMessageBox(self)
         msg.setIcon(QMessageBox.Information)
-        msg.setWindowTitle("DADOS ACEITOS")
+        msg.setWindowTitle(title)
         msg.setText(txt)
         msg.exec_()
 
@@ -188,21 +246,21 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
 
         if senha != self.le_senha2.text().strip():
-            self.messagebox_critical('As senhas informadas não são iguais\nPreencha os campos com as senhas iguais.')
+            self.messagebox_critical("Senhas Divergentes",'As senhas informadas não são iguais\nPreencha os campos com as senhas iguais.')
             self.le_senha1.setText('')
             self.le_senha2.setText('')
         elif senha == '' or self.le_senha2.text() == '' or usuario == '' or login == '':
-            self.messagebox_critical('Campos vazios\nPreencha os campos com Usuario e as senhas iguais.')
+            self.messagebox_critical("Campos vazios",'Preencha os campos com Usuario e as senhas iguais.')
             self.le_senha1.setText('')
             self.le_senha2.setText('')
         elif db.check_login_exists(login, senha):
-            self.messagebox_critical('Usuario e Senha ja Cadastrado\nPreencha os campos com um novo Usuario e senha')
+            self.messagebox_critical('Usuario e Senha ja Cadastrado','Preencha os campos com um novo Usuario e senha')
             self.le_senha1.setText('')
             self.le_senha2.setText('')
             self.le_novo_usuario.setText('')
         else:
             db.insert_novo_usuario(login,usuario,senha,acess)
-            self.messagebox_accept('Senha Cadastrada com Sucesso!')
+            self.messagebox_accept('Senha Cadastrada','Senha Cadastrada com Sucesso!')
             self.le_senha1.setText('')
             self.le_novo_login.setText('')
             self.le_novo_usuario.setText('')
@@ -216,13 +274,25 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         usuario = self.le_usuario_deletar.text()
 
         if self.le_senha_deletar.text() == self.senha and db.check_user_exists(usuario):
-            db.delete_user(usuario)
-            self.messagebox_accept('USUARIO DELETADO COM SUCESSO!')
-            self.le_usuario_deletar.setText('')
-            self.le_senha_deletar.setText('')
-            db.close_conecta()
+            msg = QMessageBox(self)
+            msg.setWindowTitle('DELETAR USUARIO?')
+            msg.setIcon(QMessageBox.Warning)
+            msg.setText(f'DELETAR O USUARIO : "{self.le_usuario_deletar.text()}"')
+            msg.setStandardButtons(QMessageBox.Yes | QMessageBox.No)
+            msg.setButtonText(QMessageBox.Yes,'SIM')
+            msg.setButtonText(QMessageBox.No,'NÃO')
+
+            r = msg.exec_()
+            if r == QMessageBox.Yes:
+                db.delete_user(usuario)
+                self.messagebox_accept('USUARIO DELETADO','USUARIO DELETADO COM SUCESSO!')
+                self.le_usuario_deletar.setText('')
+                self.le_senha_deletar.setText('')
+                db.close_conecta()
+            else:
+                pass
         else:
-            self.messagebox_critical('USUARIO NÃO CADASTRADO')
+            self.messagebox_aviso("USUARIO NÃO ENCONTRADO",'USUARIO NÃO CADASTRADO')
             self.le_usuario_deletar.setText('')
             self.le_senha_deletar.setText('')
             db.close_conecta()
@@ -232,24 +302,45 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         db.setDatabaseName('system.db')
         db.open()
 
-        self.model = QSqlTableModel(db=db)
-        self.tv_clientes.setModel(self.model)
-        self.model.setTable('clientes')
-        self.model.select()
+        self.model_clientes = QSqlTableModel(db=db)
+        self.tv_clientes.setModel(self.model_clientes)
+        self.model_clientes.setTable('clientes')
+        self.model_clientes.select()
+        self.tv_clientes.sortByColumn(0, Qt.AscendingOrder)
+        for i in range(0, 9):
+            self.tv_clientes.resizeColumnToContents(i)
 
     def show_table_estoque(self):
         db = QSqlDatabase('QSQLITE')
         db.setDatabaseName('system.db')
         db.open()
 
-        self.model = QSqlTableModel(db=db)
-        self.tv_estoque.setModel(self.model)
-        self.model.setTable('estoque')
-        self.model.select()
+        self.model_estoque = QSqlTableModel(db=db)
+        self.tv_estoque.setModel(self.model_estoque)
+        self.model_estoque.setTable('estoque')
+        self.model_estoque.select()
+        self.tv_estoque.sortByColumn(0, Qt.AscendingOrder)
+        for i in range(0, 9):
+            self.tv_estoque.resizeColumnToContents(i)
+
+    def show_table_saida(self):
+        db = QSqlDatabase('QSQLITE')
+        db.setDatabaseName('system.db')
+        db.open()
+
+        self.model_saida = QSqlTableModel(db=db)
+        self.tv_saida.setModel(self.model_saida)
+        self.model_saida.setTable('saida')
+        self.model_saida.select()
+        self.tv_saida.sortByColumn(0, Qt.AscendingOrder)
+        for i in range(0, 9):
+            self.tv_saida.resizeColumnToContents(i)
 
     def reset_tables(self):
-        self.tv_estoque.update()
         self.tv_clientes.update()
+        self.tv_estoque.update()
+        self.tv_saida.update()
+        self.show_table_saida()
         self.show_table_estoque()
         self.show_table_clientes()
 
@@ -273,10 +364,10 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         celular = self.le_celular.text().upper().strip()
 
         if nome == '':
-            self.messagebox_critical('O CAMPO "NOME" ESTA VAZIO')
+            self.messagebox_aviso("CAMPO VAZIO",'O CAMPO "NOME" ESTA VAZIO')
 
         elif nome.isnumeric():
-            self.messagebox_critical('O CAMPO "NOME" NÃO PODE CONTER NUMEROS')
+            self.messagebox_aviso('O CAMPO "NOME" NÃO PODE CONTER NUMEROS')
             self.le_nome.setText('')
